Florida's 2000 presidential election results

This database contains demographic variables and the number of votes for the major candidates, Bush and Gore, as well as the ballots cast for Pat Buchanan for each of the 67 counties.

buchanan

Format

a data frame in long format of 67 observations on the following variables

accidents

integer; daily number of accident

county

name of county

popn

population of the county in 1997

white

percentage of whites in 1996

black

percentage of blacks in 1996

hisp

percentage of Hispanics in 1996

geq65

percentage of the population aged 65 and above based on 1996 and 1997 population estimates

highsc

percentage of the population with a high school degree (1990 Census data)

coll

percentage of the population that are college graduates (1990 Census data)

income

mean personal income in 1994

buch

total ballots cast for Pat Buchanan (Reform)

bush

total ballots cast for Georges W. Bush (GOP)

gore

total ballots cast for Al Gore (Democrat)

totmb

the total number of votes cast for the presidential election in each county, minus Buchanan votes

Source

Smith, R.L. (2002). A statistical assessment of Buchanan's vote in Palm Beach County, 17(4), 441-457.