College gender inequality study

The 2008-09 nine-month academic salary for Assistant Professors, Associate Professors and Professors in a college in the U.S. The data were collected as part of the on-going effort of the college's administration to monitor salary differences between male and female faculty members.

college

Format

a data frame with 397 observations of the following variables

rank

academic rank of the professor (assistant, associate or full).

field

categorical variable indicating whether research field is applied or theoretical.

years

number of years since PhD

sex

binary indicator, zero for men (man) and unity for women (woman).

service

number of years of service in the college.

salary

nine-month income (in thousand USD).

Source

Fox J. and Weisberg, S. (2019) An R Companion to Applied Regression, Third Edition, Sage.

See also