Practical driving exam results

These records were extracted from GovUK and cleaned: they give the success rate of practical driving license exams in the United Kingdom in 2018 in 346 driving centers. A total of 761 750 people passed the test, out of 1 663 897 trials

gbdriving

Format

a data frame in long format containing the following variables, grouped by center

pass

number of successfull candidates (by sex)

total

total number of trials within the center (by sex)

region

factor indicating where the exam took place, either Scotland, Wales or one the nine regions of England

sex

factor, either man or woman

Source

https://www.gov.uk/government/statistical-data-sets/driving-test-statistics-drt