Intention to buy candies based on an advertisement

In a study performed at Tech3Lab, subjects navigated a website that contained, among other things, an advertisement for candies. During the site navigation, an ``eye-tracker'' measured the location on the screen on which the subject's eyes were fixated. The tracker also recorded whether the subject saw the ad and for how long it was in sight. Additionally, facial expression analysis software (FaceReader) can be used to guess the subject's emotions when the ad was in sight. At the end of the study, a questionnaire measured the subject's intention to buy this type of candy and socio-demographic variables. Only the 120 subjects that had seen the ad in question are included in the data

intention

Format

a data frame of 120 observations containing the following variables

intention

discrete variable ranging between 2 and 14; larger values indicate higher interest in buying the product. Specifically, the score was constructed by summing the response of two questions, both measured using a Likert scale ranging from strongly disagree (1) to strongly agree (7).

fixation

the total duration of fixation on the ad (in seconds).

emotion

a measure of reaction during fixation; the ratio of the probability of showing a positive emotion to the probability of showing a negative emotion.

sex

sex of subject, either man (0) or woman (1).

age

age (in years).

revenue

categorical variable indicating the subject's annual income; one of (1) [0, 20k); (2) [20k, 60k); (3) 60k and above.

educ

categorical variable indicating the highest educational achievement, either (1) high school or lower; (2) college or (3) university degree.

marital

civil status, either single (0) or in a relationship (1).

buy

binary variable indicating if the person bought candies.

nitem

number of candy packs bought.

Source

Marc Fredette