Renfe train tickets

A subset of a larger database containing information about train tickets sold by Renfe between Madrid and Barcelona. There are three types of trains: either Spanish high speed trains Alta Velocidad Española (AVE), or AVE-TGV which are run in partnership with SNCF to and from France (via Toulouse), or regional trains REXPRESS; only the first two labelled AVE and AVE-TGV are high speed trains.

renfe

Format

a data frame of 10k observations containing the following variables:

price

price of ticket (in euros)

dest

factor, binary indicator for destination (Barcelona-Madrid or Madrid-Barcelona)

fare

factor, categorical variable for ticket fare; one of AdultoIda, Promo or Flexible

class

factor, categorical variable for ticket class; one of Preferente, Turista, TuristaPlus or TuristaSolo

type

factor, categorical variable for train type; either AVE-TGV, AVE or REXPRESS.

duration

advertised duration of journey (in minutes)

wday

factor integer denoting week day of travel starting from Sunday(1) to Saturday (7)

Source

Kaggle https://www.kaggle.com/thegurusteam/spanish-high-speed-rail-system-ticket-pricing